requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
nats-py==2.7.2